gets items from base for index.jsp

+++ src/DataBase/AccountDB.java
@@ -294,7 +294,7 @@ public class AccountDB implements DBQueries {
             try (ResultSet rs = stm.executeQuery()) {
                 if (rs.next()) {
                     item = ObjectFactory.getNewItem(rs.getString("ItemName"), rs.getInt("itemID"), rs.getInt("ownerID"),
-                            rs.getString("itemImageUrl"), rs.getFloat("price"), rs.getInt("categoryID"), rs.getInt("rating"),
+                            rs.getString("itemImageUrl"), rs.getDouble("price"), rs.getInt("categoryID"), rs.getInt("rating"),
                             rs.getInt("voters"), rs.getString("description"));
                 }
             } catch (SQLException e) {
@@ -323,7 +323,7 @@ public class AccountDB implements DBQueries {
             try (ResultSet rs = stm.executeQuery()) {
                 while (rs.next()) {
                     ls.add(ObjectFactory.getNewItem(rs.getString("ItemName"), rs.getInt("itemID"), rs.getInt("ownerID"),
-                            rs.getString("itemImageUrl"), rs.getFloat("price"), rs.getInt("categoryID"), rs.getInt("rating"),
+                            rs.getString("itemImageUrl"), rs.getDouble("price"), rs.getInt("categoryID"), rs.getInt("rating"),
                             rs.getInt("voters"), rs.getString("description")));
                 }
             } catch (SQLException e) {
@@ -341,7 +341,7 @@ public class AccountDB implements DBQueries {
             try (ResultSet rs = stm.executeQuery()) {
                 if (rs.next()) {
                     return ObjectFactory.getNewItem(rs.getString("ItemName"), rs.getInt("itemID"), rs.getInt("ownerID"),
-                            rs.getString("itemImageUrl"), rs.getFloat("price"), rs.getInt("categoryID"), rs.getInt("rating"),
+                            rs.getString("itemImageUrl"), rs.getDouble("price"), rs.getInt("categoryID"), rs.getInt("rating"),
                             rs.getInt("voters"), rs.getString("description"));
                 }
             } catch (SQLException e) {
